# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Nodejs - Todos - more middlewares | Documentation",
    "description": "SOLID challenge API Documentation",
    "version": "1.0.0"
  },
  "paths": {
    "/users": {
      "post": {
        "tags": ["Users"],
        "summary": "Create user",
        "description": "Create a new user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "username": {
                    "type": "string"
                  }
                },
                "example": {
                  "name": "contoso name",
                  "username": "c_name"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Username already exists"
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": ["Users"],
        "summary": "Show an user",
        "description": "List a specific user",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id that need to be list",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "Please check the id again"
          }
        }
      }
    },
    "/users/{id}/pro": {
      "patch": {
        "tags": ["Users"],
        "summary": "Update user's plan",
        "description": "Activate user's plan to pro",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id that need to be list",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Pro plan is already activated."
          }
        }
      }
    },
    "/todos": {
      "post": {
        "tags": ["Todos"],
        "summary": "Create todo",
        "description": "Create a new todo",
        "parameters": [
          {
            "name": "username",
            "in": "header",
            "description": "username that need to be list",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "deadline": {
                    "type": "string"
                  }
                },
                "example": {
                  "title": "A new Title",
                  "deadline": "Fri Nov 12 2021 18:00:55 GMT-0300 (Brasilia Standard Time)"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "403": {
            "description": "You already have ten todos"
          },
          "404": {
            "description": "This username does not exists!"
          }
        }
      },
      "get": {
        "tags": ["Todos"],
        "summary": "Show todos",
        "description": "List all todos from give an username",
        "parameters": [
          {
            "name": "username",
            "in": "header",
            "description": "username that need to be list",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Please check the username again"
          }
        }
      }
    },
    "/todos/{id}": {
      "put": {
        "tags": ["Todos"],
        "summary": "Update todo",
        "description": "Update an existent todo",
        "parameters": [
          {
            "name": "username",
            "in": "header",
            "description": "username that need to be list",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "id that need to be list",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "deadline": {
                    "type": "string"
                  }
                },
                "example": {
                  "title": "A title updated",
                  "deadline": "Fri Nov 12 2021 18:00:55 GMT-0300 (Brasilia Standard Time)"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "The Id is not a uuid"
          },
          "404": {
            "description": "This user does not exists or The id is not a uuid"
          }
        }
      },
      "delete": {
        "tags": ["Todos"],
        "summary": "Delete todo",
        "description": "Delete an existent todo",
        "parameters": [
          {
            "name": "username",
            "in": "header",
            "description": "username that need to be list",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "id that need to be list",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          },
          "404": {
            "description": "This todo does not exists!"
          }
        }
      }
    },
    "/todos/{id}/done": {
      "patch": {
        "tags": ["Todos"],
        "summary": "Done a todo",
        "description": "Update an existent todo to done",
        "parameters": [
          {
            "name": "username",
            "in": "header",
            "description": "username that need to be list",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "id that need to be list",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "This todo does not exists!"
          }
        }
      }
    }
  }
}